Implement cpu pinning and cgroup cpuset support and add mixxx-isolate helper script.
Add command line argument to specify a cpuid and cpuset
cgroup name for the engine.
The engine tries at startup to move itself to the cpuset
provided or set the cpu affinity.
This will allow a wrapper program to free a core for the engine,
minimizing latency jitter or CPU starvating.

Add script which will isolate a dedicated CPU for the mixxx engine.

- Select a CPU and its hyperthread partners (last real cpu)
- Create cpuset groups for the system with all other cpus and move all
processes there.
- create a cpuset group for the engine and one for all hyperthreading partners
- try to take hyperthread cpu offline
- start mixx with provided paramenters and cgroup parameters

when mixxx is stopped
- take cpu online again
- move all processes back to the root group
- delete cpuset groups